Don't crash with invalid axes.
If the publication of the Joy message didn't contain enough
axes, the code would previously access beyond the bounds of
the vector and crash.  This changes it so that we check to
ensure that the incoming message has enough axes to satisfy
our constraints.  While we are at it, consolidate all of the
very similar code together so it is all in one place.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>